Moved DataType into it's own class (resolved a name clash on ARDOUR::MIDI which was a show stopper for gcc 3.3) Also fixed a few warnings in fft_graph.cc
git-svn-id: svn://localhost/ardour2/trunk@739 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ gtk2_ardour/fft_graph.cc
@@ -234,11 +234,11 @@ FFTGraph::draw_scales(Glib::RefPtr<Gdk::Window> window)
 		while (_logScale[logscale_pos] < position_on_scale)
 			logscale_pos++;
 		
-		int coord = v_margin + 1.0 + position_on_scale;
+		int coord = (int)(v_margin + 1.0 + position_on_scale);
 		
 		int SR = 44100;
 
-		int rate_at_pos = (double)(SR/2) * (double)logscale_pos / (double)_dataSize;
+		int rate_at_pos = (int)((double)(SR/2) * (double)logscale_pos / (double)_dataSize);
 		
 		char buf[32];
 		snprintf(buf,32,"%dhz",rate_at_pos);
